+++ webpack.config.js
@@ -39,6 +39,7 @@ module.exports = {
               presets: ['@babel/preset-env','@babel/react'],
               "plugins": [
                 "@babel/plugin-proposal-class-properties",
+                "@babel/plugin-syntax-dynamic-import",
                 [
                   "create-automation-page-objects",
                   {