+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Beta User Acceptance Testing]
 - IdentityProviders
 
+## [0.97.0] - 2021-02-01
+### Added
+- CXV1-23250 - Add pdf export functionality to forecast dashboards.
+- CXV1-23709 - Add export to csv functionality to forecasting dashboards.
+
 # [0.96.6] _ 2021-01-25
 ### Fixed
 - <no jira> - Fix Automation Issue for random failing of test.